fix as per comment
